chore(connect-data): move firmware public key from env-utils
