+++ packages/connect-common/tsconfig.json
@@ -6,7 +6,6 @@
     },
     "include": ["."],
     "references": [
-        { "path": "../env-utils" },
         { "path": "../utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" }
