+++ packages/connect-common/tsconfig.lib.json
@@ -8,9 +8,6 @@
     "include": ["./src"],
     "references": [
         {
-            "path": "../env-utils"
-        },
-        {
             "path": "../utils"
         },
         {
