+++ packages/connect-data/src/index.ts
@@ -1,1 +1,2 @@
+export * from './firmware-jws';
 export * from './map-releases';
