+++ packages/connect/tsconfig.json
@@ -13,7 +13,6 @@
         { "path": "../crypto-utils" },
         { "path": "../device-authenticity" },
         { "path": "../device-utils" },
-        { "path": "../env-utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" },
         { "path": "../schema-utils" },
