+++ packages/connect/tsconfig.lib.json
@@ -32,9 +32,6 @@
             "path": "../device-utils"
         },
         {
-            "path": "../env-utils"
-        },
-        {
             "path": "../protobuf"
         },
         {
